@@ -2,7 +2,7 @@
 name = "gst_pipeline_studio"
 version = "0.2.3"
 edition = "2018"
-rust-version = "1.63"
+rust-version = "1.67"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
